Moving memory services into services/memory/ dir.
